requirements.txt
Pillow>=10.0.0
dashscope>=1.20.0
zhipuai>=2.1.0
pytest>=9.0
pytest-asyncio>=0.23

# --- Server dependencies ---
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
pymilvus>=2.4
boto3>=1.35
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
